feat: render correct form on edit
